Add toolbar button hover and click effects for legacy parity (#252)
