+++ .gitignore
@@ -312,8 +312,6 @@ copilot-chat/**
 .claude-code/**
 .anthropic/
 .anthropic/**
-.agentbridge/
-.agentbridge/**
 
 # --- Cursor ---
 .cursor/
