+++ src/SkyCD.App/Models/AppOptions.cs
@@ -29,6 +29,4 @@ public sealed class AppOptions
     public List<string> DisabledPluginIds { get; set; } = [];
 
     public int OptionsTabIndex { get; set; }
-
-    public string? LastOpenedCatalogPath { get; set; }
 }
