+++ src/SkyCD.App/Styles/ButtonStyles.axaml
@@ -1,15 +1,17 @@
 <Styles xmlns="https://github.com/avaloniaui"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
-    <Style Selector="Button:pressed">
-        <Setter Property="Opacity" Value="0.7"/>
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:cc="clr-namespace:SkyCD.UI.Controls;assembly=SkyCD.UI">
+    <Style Selector="cc|ClassicToolbarButton">
+        <Setter Property="Background" Value="Transparent"/>
+        <Setter Property="Padding" Value="2,1"/>
     </Style>
-    <Style Selector="Button:pointerover">
-        <Setter Property="Opacity" Value="0.9"/>
+    <Style Selector="cc|ClassicToolbarButton:pointerover">
+        <Setter Property="Background" Value="{DynamicResource ThemeBorderMidBrush}"/>
     </Style>
-    <Style Selector="Button:pointerover /template/ ContentPresenter">
-        <Setter Property="Background" Value="#40FFFFFF"/>
+    <Style Selector="cc|ClassicToolbarButton:pressed">
+        <Setter Property="Background" Value="{DynamicResource ThemeBorderHighBrush}"/>
     </Style>
-    <Style Selector="Button:pressed /template/ ContentPresenter">
-        <Setter Property="Background" Value="#60FFFFFF"/>
+    <Style Selector="cc|ClassicToolbarButton:disabled">
+        <Setter Property="Opacity" Value="0.5"/>
     </Style>
 </Styles>
